Move Main into a jp package
And use exec in the bin script
