+++ bin/jp
@@ -3,4 +3,4 @@
 classpath=$(find target/dist/lib -name '*.jar' -exec echo $(pwd)/{} \; | tr '\n' ':')
 classpath="$classpath:target/classes"
 
-java -classpath "$classpath" io.burt.jmespath.Main "$@"
+exec java -classpath "$classpath" io.burt.jmespath.jp.Main "$@"
